requirements.txt
discord.py
python-dotenv
nltk
openai
psycopg2-binary